# Conflicts:
#	Arknight/Assets/Scripts/MainScene/Tower/TowerManager.cs

+++ Arknight/Assets/Scripts/MainScene/Tower/TowerManager.cs
@@ -18,13 +18,13 @@ public class TowerManager : MonoBehaviour
         IDLE, BATTLE, DEATH               // 대기, 전투, 사망
     }
     public STATE m_State;            // 상태 받는 변수
-    public int m_MaxHp;              // 최대 체력
+    public float m_MaxHp;              // 최대 체력
     public float m_CurrentHp;                 // 체력
-    public int m_MaxMp;              // 최대 마력
-    public int m_CurrentMp;                 // 마력   
+    public float m_MaxMp;              // 최대 마력
+    public float m_CurrentMp;                 // 마력   
     public int m_TileX;              // 타워 X좌표
     public int m_TileY;              // 타워 Y좌표
-    public int m_Damage;             // 공격력
+    public float m_Damage;             // 공격력
     public float m_AttackDelay;      // 공격 딜레이
     public float m_AttackDist;       // 사거리
